Compile Python with UCS-4.
Twitter entities indices are calculated with the assumption that Unicode
characters outside the basic multilingual plane are 4 bytes wide (UCS-4), i.e.
not variable-length encoded (UCS-2). Thus, we must have a Python that
internally represents Unicode as UCS-4. Unfortunately, that is not the default
at compile time and a switch is needed (--enable-unicode=ucs4) to use UCS-4.