Fix Twitter streaming API changes.
Twitter has moved all of its streaming APIs to HTTP 1.1 chunked encoded
transfer and broke the python library we use. This change fixes this. We use a
branch from a user who fixed the library, but the change has not been ported
back into it. The main developer is not responding and it seems we're looking
at a fork of this code going forward.